Multiple changes for this commit
Ensembl

Code for worm used an archived version of a Biomart database that was supposed to be
backed out for the next release but apparently never was. In addition, the biomaRt code
used a non-existent (as of now) attribute column of entrezegene whereas it is really entrezegene_id.

The call to biomaRt for chipmapsrc_worm.rsqlite adds in a bunch of lines to the wormbase table, most of which appear to be
dups of existing data. Added SQL code to the bottom of bindb.R to remove any duplicate rows.

Gene

Added a test at the end of getdb.sh, which calls an R script to go through all the chipmapsrc_xxx.sqlite files
that were just generated and test for any duplicated lines in any of the tables.

GO

Finally got this fixed. The issue has to do with the fact that GO thinks of the DAG as being
a directed graph, going from child -> parent. But for our purposes, we want the directionality to be
reversed. In other words, we generate a set of tables that have all the offspring for a given GO term,
in which case we need first to have the parent -> child directed graph, and then to do a transitive closure
to get all the offspring for each GO term (e.g., for each term we want all the offspring GO terms that
are reachable from that term). Fixed parseOBO.R to switch the direction of the directed graph prior to running
transitive_closure. Also, the term2term.txt file being output has term1 (child) and term2 (parent) in oclumns 4 and 5.
The code in bindb4.sql incorrectly input those data in the go_parents table, putting term2 in as the ID and term1 in as the value,
which is wrong, since the go_parents table should have term1 (the child) as the _id and term2 (the parent) as the table value.