Several additions for this commit.
Converted makeTerminalDbs.R to an Rscript format, so it can be called once to
build the OrgDb, GO.db and PFAM.db packages, and again to build the TxDb packages
without having to comment or uncomment text.

Added two new R scripts, regenMapCounts.R and testDbs.R. The former is intended
to regenerate the map_counts.sqlite DB that is supposed to be in annosrc/db, in case
it gets deleted. The README currently says to only keep the metadatasrc.sqlite, but
that is incorrect. We use the map_counts.sqlite DB to compare the counts/table in
the newly built DBs to ensure they aren't too different from the last build. If it
accidently gets deleted, regenMapCounts.R will be called, and will extract the map_counts
data from the installed db0 packages (and GO.db, etc) to regenerate the map_counts DB with
the data from the last build.

The code in testDbs.R is intended to find tables in any of the newly created DBs that have
missing data. For example, there are 3900 missing values in the flybase_prot table of
chipsrc_fly.sqlite, where we have an _id, but we don't have a prot_id. Both of these scripts
get called as part of map_counts/scripts/getdb.sh, which checks for aberrant changes in the new
DBs as compared to the old ones.

Additionally, changes to anosrc/inparanoid/script/bindb_flyExtras.sql, which was
the cause of the problems noted in the previous paragraph.