Two changes for this commit
Changes to the new DB checking script, where the test is for any rows
that have '' for all entries except for the _id column.

Changes to makeTerminalDBPkgs.R. This script by default uses a text file
that is in the scripts subdirectory of the GenomicsFeatures package. I get
why that sort of kludge is necessary for something like the OrgDb packages,
where there are so many DB types and we make them all each time.

But for the TxDb packages, where we build them only if they get changed at
UCSC, I can't see why we need that sort of setup. Plus it's now necessary to
pass in a correctly formed circ_seqs argument for makeTxDbPackageFromUCSC, and
it's difficult to do that using the current paradigm.

Instead, there are now two character vectors in makeTerminalDbs.R that need to be
adjusted for each release. The first has all the genomes that need to be built, and
the second has the requisite table to use. The circ_seqs argument is then generated
on the fly, for the genomes that are going to be used.