This is a fix for generating GO.db. Previously we downloaded a set of flat files from geneontology.org that we are now trying to replicate by parsing the OBO files (the flat files are no longer updated). There is a virtual node, called 'all' that is the parent node to the top node for all three ontologies, which was included in the flat files, but is not parsed out of the OBO.
This term crops up in the term2term.txt file, which lists all the parent -> child
mappings, and also in the graph_path.txt file that lists all the transitive closures
for each term.

I changed the parseOBO.R script to recapitulate as closely as possible the structure
of the flat files from geneontology.org, including the 'all' mappings. I also reversed
the mapping directions (which were correct for the format of the files I generated last time,
but were not correct for the consistentized flat files).